+++ llama_index/indices/service_context.py
@@ -1,4 +1,3 @@
-import dataclasses
 import logging
 from dataclasses import dataclass
 from typing import Optional
@@ -39,9 +38,9 @@ def _get_default_prompt_helper(
 ) -> PromptHelper:
     """Get default prompt helper."""
     if context_window is not None:
-        llm_metadata = dataclasses.replace(llm_metadata, context_window=context_window)
+        llm_metadata.context_window = context_window
     if num_output is not None:
-        llm_metadata = dataclasses.replace(llm_metadata, num_output=num_output)
+        llm_metadata.num_output = num_output
     return PromptHelper.from_llm_metadata(llm_metadata=llm_metadata)