+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bug Fixes / Nits
 - Added streaming support to simple and condense chat engines (#6717)
 - fixed `response_mode="no_text"` response synthesizer (#6755)
+- fixed error setting `num_output` and `context_window` in service context (#6766)
 
 ## [v0.7.2] - 2023-07-06